Add methods to create and delete lambda functions.
Deletion was a little odd, since function-name has to be passed in
the body of the request, but is actually used in the url. And method
always has to be passed as :delete.

These could be wrapped in the utils namespace to make them easier to use,
but it's possibly better to alter their use in the core lambda ns.

The base64 string in the test ns was created from the test.js file in the
project here: https://github.com/lfn3/node-cljs-zip